requirements.txt
pytest>=8.0.0
Flask>=3.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.8.0
langchain>=0.1.0
ollama>=0.1.0
numpy>=1.23.0
scipy>=1.10.0
pytest-cov>=4.0.0
openai>=1.0.0
